+++ tower-util/src/layer/mod.rs
@@ -1,0 +1,5 @@
+mod chain;
+mod identity;
+
+pub use self::chain::Chain;
+pub use self::identity::Identity;
