+++ tower-util/src/lib.rs
@@ -7,11 +7,13 @@ extern crate futures;
 #[cfg(feature = "io")]
 extern crate tokio_io;
 extern crate tower_service;
+extern crate tower_layer;
 
 mod boxed;
 mod call_all;
 #[cfg(feature = "either")]
 mod either;
+pub mod layer;
 #[cfg(feature = "io")]
 mod make_connection;
 mod make_service;
